+++ tests/urls.py
@@ -4,9 +4,6 @@ from django.contrib import admin
 import nexus
 from nexus.compat import subinclude
 
-admin.autodiscover()
-nexus.autodiscover()
-
 urlpatterns = [
     url(r'^admin/', subinclude(admin.site.urls)),
     url(r'^nexus/', include(nexus.site.urls)),